feat: Add Export Initial and Progress State components
- Introduced `ExportInitialState.tsx` to manage the initial export selection UI, including bulk selection and content options.
- Added `ExportProgressState.tsx` to display the progress of the export process with a progress bar and status messages.
- Updated `ExportModal.tsx` to integrate the new components and streamline the export workflow.
- Removed unused utility functions related to formatting completed items.
